+++ src/assets/scripts/aircraft/Leg.js
@@ -1,233 +1,339 @@
+import _forEach from 'lodash/forEach';
+import _get from 'lodash/get';
 import _has from 'lodash/has';
 import _map from 'lodash/map';
 import Waypoint from './Waypoint';
+import RouteModel from '../airport/Route/RouteModel';
+import { FP_LEG_TYPE } from '../constants/aircraftConstants';
 import { LOG } from '../constants/logLevel';
 
-// can be 'sid', 'star', 'iap', 'awy', 'fix', '[manual]'
 /**
- * Enumeration of possibl FLight Plan Leg types.
- *
- * @property FP_LEG_TYPE
- * @type {Object}
- * @final
- */
-export const FP_LEG_TYPE = {
-    SID: 'sid',
-    STAR: 'star',
-    IAP: 'iap',
-    AWY: 'awy',
-    FIX: 'fix',
-    MANUAL: '[manual]'
-};
-
-/**
-  * Build a 'leg' of the route (contains series of waypoints)
+  * This class acts as a collection of Waypoint model objects.
   *
-  * @param {object} data = {route: "KSFO.OFFSH9.SXC", either a fix, or with format 'start.procedure.end', or
-  *                                                   "[RNAV/GPS]" for custom positions
-  *                         type: "sid",              can be 'sid', 'star', 'iap', 'awy', 'fix'
-  *                         firstIndex: 0}            the position in fms.legs to insert this leg
+  * @class Leg
   */
 export default class Leg {
     /**
+     *
+     * @for Leg
      * @constructor
+     * @param data
+     * @param {object} route:           "KSFO.OFFSH9.SXC", either a fix, or with format 'start.procedure.end', or
+     *                                  "[RNAV/GPS]" for custom positions
+     *                 type: "sid",     can be 'sid', 'star', 'iap', 'awy', 'fix'
+     *                 firstIndex: 0    the position (index) in fms.legs to insert this leg
      */
     constructor(data = {}, fms) {
-        this.route = '[radar vectors]'; // eg 'KSFO.OFFSH9.SXC' or 'FAITH'
-        this.type = FP_LEG_TYPE.MANUAL;
-        this.waypoints = []; // an array of zlsa.atc.Waypoint objects to follow
+        /**
+         * String representation of a route.
+         *
+         * May be a single fix or a route expressed in dot notation. ex:
+         * - `KSFO.OFFSH9.SXC`
+         * - `FAITH`
+         *
+         * @property route
+         * @type {string}
+         * @default ''
+         */
+        this.route = '';
 
-        // Fill data with default Leg properties if they aren't specified (prevents wp constructor from getting confused)
-        if (!data.route) {
-            data.route = this.route;
-        }
+        /**
+         * @property type
+         * @type {string}
+         * @default ''
+         */
+        this.type = '';
 
-        if (!data.type) {
-            data.type = this.type;
-        }
-
-        if (!data.waypoints) {
-            data.waypoints = this.waypoints;
-        }
+        // TODO: possibly implement as a waypointCollection
+        /**
+         * A collection of Waypoint instances
+         *
+         * @property waypoints
+         * @type {Array}
+         * @default []
+         */
+        this.waypoints = [];
 
         this.parse(data, fms);
     }
 
     /**
      * Parse input data and apply to this leg
+     *
+     * @for Leg
+     * @method parse
+     * @param data {object}
+     * @param fms {AircraftFlightManagementSystem}
      */
     parse(data, fms) {
-        // FIXME: make these multi-line, single lines are hard to reason and prone to errors.
-        for (const i in data) if (this.hasOwnProperty(i)) this[i] = data[i]; // Populate Leg with data
-        if (this.waypoints.length === 0) this.generateWaypoints(data, fms);
-            if (this.waypoints.length === 0) this.waypoints = [new Waypoint({ route: '' }, fms)];
+        // TODO: move radar vectors to constants file
+        this.route = _get(data, 'route', '[radar vectors]');
+        this.type = _get(data, 'type', FP_LEG_TYPE.MANUAL);
+        this.waypoints = _get(data, 'waypoints', []);
+
+        if (this.waypoints.length === 0) {
+            this.generateWaypoints(data, fms);
+        }
     }
 
     /**
-     * Adds Waypoint objects to this Leg, based on the route & type
+     * Adds Waypoint objects to this Leg based on the route type
+     *
+     * @for Leg
+     * @method generateWaypoints
+     * @param data {object}
+     * @param fms {AircraftFlightManagementSystem}
      */
     generateWaypoints(data, fms) {
         if (!this.type) {
             return;
         }
 
-        if (this.type === FP_LEG_TYPE.SID) {
-            if (!fms) {
-                log('Attempted to generate waypoints for SID, but cannot because fms ref not passed!', LOG.WARNING);
+        const airport = window.airportController.airport_get();
 
-                return;
-            }
+        switch (this.type) {
+            case FP_LEG_TYPE.SID:
+                // TODO: this is gross. we instantiate route with a string and new mutate it here to a RouteModel.
+                this.route = new RouteModel(data.route);
+                this._generateWaypointsForSid(data, fms);
 
-            // const { apt, sid, exit } = data.route.split('.');
-            const apt = data.route.split('.')[0];
-            const sid = data.route.split('.')[1];
-            const exit = data.route.split('.')[2];
-            const rwy = fms.my_aircraft.rwy_dep;
-            this.waypoints = [];
+                break;
+            case FP_LEG_TYPE.STAR:
+                // TODO: this is gross. we instantiate route with a string and new mutate it here to a RouteModel.
+                this.route = new RouteModel(data.route);
+                this._generateWaypointsForStar(data, fms);
 
-            // Generate the waypoints
-            if (!rwy) {
-                const isWarning = true;
-                window.uiController.ui_log(`${fms.my_aircraft.getCallsign()} unable to fly SID, we haven't been assigned a departure runway!`, isWarning);
+                break;
+            case FP_LEG_TYPE.IAP:
+                // FUTURE FUNCTIONALITY
+                this._generateWaypointsForIap(data, airport);
 
-                return;
-            }
+                break;
+            case FP_LEG_TYPE.AWY:
+                // TODO: this is gross. we instantiate route with a string and new mutate it here to a RouteModel.
+                this.route = new RouteModel(data.route);
+                this._generateWaypointsForAirway(data, airport);
 
-            const pairs = window.airportController.airport_get(apt).getSID(sid, exit, rwy);
+                break;
+            case FP_LEG_TYPE.FIX:
+                this._generateWaypointForFix(airport);
 
-            // Remove the placeholder leg (if present)
-            if (fms.my_aircraft.wow() && fms.legs.length > 0
-                && fms.legs[0].route === window.airportController.airport_get().icao && pairs.length > 0
-            ) {
-                // remove the placeholder leg, to be replaced below with SID Leg
-                fms.legs.splice(0, 1);
-            }
+                break;
+            case FP_LEG_TYPE.MANUAL:
+                this._generateManualWaypoint(airport);
 
-            // for each fix/restr pair
-            for (let i = 0; i < pairs.length; i++) {
-                const f = pairs[i][0];
-                let a = null;
-                let s = null;
+                break;
+            default:
+                this._generateEmptyWaypoint(airport);
 
-                if (pairs[i][1]) {
-                    const a_n_s = pairs[i][1].toUpperCase().split('|');
-
-                    for (const j in a_n_s) {
-                        if (a_n_s[j][0] === 'A') {
-                            a = a_n_s[j].substr(1);
-                        } else if (a_n_s[j][0] === 'S') {
-                            s = a_n_s[j].substr(1);
-                        }
-                    }
-                }
-
-                this.waypoints.push(new Waypoint(
-                    {
-                        fix: f,
-                        fixRestrictions: {
-                            alt: a,
-                            spd: s
-                        }
-                    },
-                    fms
-                ));
-            }
-
-            if (!this.waypoints[0].speed) {
-                this.waypoints[0].speed = fms.my_aircraft.model.speed.cruise;
-            }
-        } else if (this.type === FP_LEG_TYPE.STAR) {
-            if (!fms) {
-                log('Attempted to generate waypoints for STAR, but cannot because fms ref not passed!', LOG.WARNING);
-
-                return;
-            }
-
-            const entry = data.route.split('.')[0];
-            const star = data.route.split('.')[1];
-            const apt = data.route.split('.')[2];
-            const rwy = fms.my_aircraft.rwy_arr;
-            this.waypoints = [];
-
-            // Generate the waypoints
-            const pairs = window.airportController.airport_get(apt).getSTAR(star, entry, rwy);
-
-            // for each fix/restr pair
-            for (let i = 0; i < pairs.length; i++) {
-                const f = pairs[i][0];
-                let a = null;
-                let s = null;
-
-                if (pairs[i][1]) {
-                    const a_n_s = pairs[i][1].toUpperCase().split('|');
-
-                    for (const j in a_n_s) {
-                        if (a_n_s[j][0] === 'A') {
-                            a = a_n_s[j].substr(1);
-                        } else if (a_n_s[j][0] === 'S') {
-                            s = a_n_s[j].substr(1);
-                        }
-                    }
-                }
-
-                this.waypoints.push(new Waypoint(
-                    {
-                        fix: f,
-                        fixRestrictions: {
-                            alt: a,
-                            spd: s
-                        }
-                    },
-                    fms
-                ));
-            }
-
-            if (!this.waypoints[0].speed) {
-                this.waypoints[0].speed = fms.my_aircraft.model.speed.cruise;
-            }
-        } else if (this.type === FP_LEG_TYPE.IAP) {
-            // FUTURE FUNCTIONALITY
-        } else if (this.type === FP_LEG_TYPE.AWY) {
-            const start = data.route.split('.')[0];
-            const airway = data.route.split('.')[1];
-            const end = data.route.split('.')[2];
-            // Verify airway is valid
-            const apt = window.airportController.airport_get();
-
-            if (!_has(apt, 'airways') || !_has(apt.airways, 'airway')) {
-                log(`Airway ${airway} not defined at ${apt.icao}`, LOG.WARNING);
-                return;
-            }
-
-            // Verify start/end points are along airway
-            const awy = apt.airways[airway];
-            if (!(awy.indexOf(start) !== -1 && awy.indexOf(end) !== -1)) {
-                log(`Unable to follow ${airway} from ${start} to ${end}`, LOG.WARNING);
-                return;
-            }
-
-            // Build list of fixes, depending on direction traveling along airway
-            const fixes = [];
-            const readFwd = (awy.indexOf(end) > awy.indexOf(start));
-
-            if (readFwd) {
-                for (let f = awy.indexOf(start); f <= awy.indexOf(end); f++) {
-                    fixes.push(awy[f]);
-                }
-            } else {
-                for (let f = awy.indexOf(start); f >= awy.indexOf(end); f--) {
-                    fixes.push(awy[f]);
-                }
-            }
-
-            // Add list of fixes to this.waypoints
-            this.waypoints = [];
-            this.waypoints = _map(fixes, (f) => new Waypoint({ fix: f }, fms));
-        } else if (this.type === FP_LEG_TYPE.FIX) {
-            this.waypoints = [];
-            this.waypoints.push(new Waypoint({ fix: data.route }, fms));
-        } else {
-            this.waypoints.push(new Waypoint(data, fms));
+                break;
         }
+    }
+
+    /**
+     * Add a new Waypoint to the collection
+     *
+     * @method addWaypointToLeg
+     * @param waypointToAdd {Waypoint}
+     */
+    addWaypointToLeg(waypointToAdd) {
+        if (!(waypointToAdd instanceof Waypoint)) {
+            throw new TypeError('Invalid parameter, expecte waypointToAdd to be an instanceof the Waypoint class');
+        }
+
+        this.waypoints.push(waypointToAdd);
+    }
+
+    /**
+     * @for Leg
+     * @method _generateWaypointsForSid
+     * @param data {object}
+     * @param fms {AircraftFlightManagementSystem}
+     * @private
+     */
+    _generateWaypointsForSid(data, fms) {
+        if (!fms) {
+            log('Attempted to generate waypoints for SID, but cannot because fms ref not passed!', LOG.WARNING);
+
+            return;
+        }
+
+        this._resetWaypoints();
+
+        const rwy = fms.my_aircraft.rwy_dep;
+
+        if (!rwy) {
+            const isWarning = true;
+
+            window.uiController.ui_log(
+                `${fms.my_aircraft.getCallsign()} unable to fly SID, we haven't been assigned a departure runway!`,
+                isWarning
+            );
+
+            return;
+        }
+
+        const airport = window.airportController.airport_get(this.route.entry);
+        const waypointsForSid = airport.findWaypointModelsForSid(this.route.procedure, rwy, this.route.exit);
+
+        // TODO: refactor/abstract this boolean logic
+        // Remove the placeholder leg (if present)
+        if (fms.my_aircraft.wow() &&
+            fms.legs.length > 0 &&
+            fms.legs[0].route === airport.icao &&
+            pairs.length > 0
+        ) {
+            // remove the placeholder leg, to be replaced below with SID Leg
+            fms.legs.splice(0, 1);
+        }
+
+        for (let i = 0; i < waypointsForSid.length; i++) {
+            const waypointToAdd = waypointsForSid[i].generateFmsWaypoint(airport);
+
+            this.addWaypointToLeg(waypointToAdd);
+        }
+
+        if (!this.waypoints[0].speed) {
+            this.waypoints[0].speed = fms.my_aircraft.model.speed.cruise;
+        }
+    }
+
+    /**
+     * @for Leg
+     * @method _generateWaypointsForStar
+     * @param data {object}
+     * @param fms {AircraftFlightManagementSystem}
+     * @private
+     */
+    _generateWaypointsForStar(data, fms) {
+        if (!fms) {
+            log('Attempted to generate waypoints for STAR, but cannot because fms ref not passed!', LOG.WARNING);
+
+            return;
+        }
+
+        this._resetWaypoints();
+
+        const rwy = fms.my_aircraft.rwy_arr;
+        const airport = window.airportController.airport_get(this.route.exit);
+        const waypointsForStar = airport.findWaypointModelsForStar(this.route.procedure, this.route.entry, rwy);
+
+        for (let i = 0; i < waypointsForStar.length; i++) {
+            const waypointToAdd = waypointsForStar[i].generateFmsWaypoint(airport);
+
+            this.addWaypointToLeg(waypointToAdd);
+        }
+
+        if (!this.waypoints[0].speed) {
+            this.waypoints[0].speed = fms.my_aircraft.model.speed.cruise;
+        }
+    }
+
+    // NOT IN USE
+    _generateWaypointsForIap(data, airport) {
+        return;
+    }
+
+    // NOT IN USE
+    /**
+     * @for Leg
+     * @method _generateWaypointsForAirway
+     * @param data {object}
+     * @param fms {AircraftFlightManagementSystem}
+     * @private
+     */
+    _generateWaypointsForAirway(data, airport) {
+        const start = this.route.split('.')[0];
+        const airway = this.route.split('.')[1];
+        const end = this.route.split('.')[2];
+        // Verify airway is valid
+        const apt = window.airportController.airport_get();
+
+        if (!_has(apt, 'airways') || !_has(apt.airways, 'airway')) {
+            log(`Airway ${airway} not defined at ${apt.icao}`, LOG.WARNING);
+            return;
+        }
+
+        // Verify start/end points are along airway
+        const awy = apt.airways[airway];
+        if (!(awy.indexOf(start) !== -1 && awy.indexOf(end) !== -1)) {
+            log(`Unable to follow ${airway} from ${start} to ${end}`, LOG.WARNING);
+            return;
+        }
+
+        // TODO: abstract this logic
+        // Build list of fixes, depending on direction traveling along airway
+        const fixes = [];
+        const readFwd = (awy.indexOf(end) > awy.indexOf(start));
+
+        if (readFwd) {
+            for (let f = awy.indexOf(start); f <= awy.indexOf(end); f++) {
+                fixes.push(awy[f]);
+            }
+        } else {
+            for (let f = awy.indexOf(start); f >= awy.indexOf(end); f--) {
+                fixes.push(awy[f]);
+            }
+        }
+
+        this._resetWaypoints();
+
+        _forEach(fixes, (fix) => {
+            const waypointToAdd = new Waypoint({ fix }, airport);
+
+            this.addWaypointToLeg(waypointToAdd);
+        });
+    }
+
+    /**
+     * @for Leg
+     * @method _generateWaypointForFix
+     * @param airport {AirportInstanceModel}
+     * @private
+     */
+    _generateWaypointForFix(airport) {
+        this._resetWaypoints();
+
+        const waypointToAdd = new Waypoint({ fix: this.route }, airport);
+
+        this.addWaypointToLeg(waypointToAdd);
+    }
+
+    /**
+     * @for Leg
+     * @method _generateManualWaypoint
+     * @param airport {AirportInstanceModel}
+     * @private
+     */
+    _generateManualWaypoint(airport) {
+        const waypointToAdd = new Waypoint({ route: this.route }, airport);
+
+        this.addWaypointToLeg(waypointToAdd);
+    }
+
+    /**
+     * @for Leg
+     * @method _generateEmptyWaypoint
+     * @param airport {AirportInstanceModel}
+     * @private
+     */
+    _generateEmptyWaypoint(airport) {
+        const waypointToAdd = new Waypoint({ route: '' }, airport);
+
+        this.addWaypointToLeg(waypointToAdd);
+    }
+
+    /**
+     * Reset the waypoint property to an empty array.
+     *
+     * Provides a single method that encapsulates common functionality that
+     * can be used throughout the class.
+     *
+     * @for Leg
+     * @method _resetWaypoints
+     * @private
+     */
+    _resetWaypoints() {
+        this.waypoints = [];
     }
 }