+++ test/airport/fix/FixCollection.spec.js
@@ -1,4 +1,5 @@
 import ava from 'ava';
+import _isEqual from 'lodash/isEqual';
 
 import FixCollection from '../../../src/assets/scripts/airport/Fix/FixCollection';
 import FixModel from '../../../src/assets/scripts/airport/Fix/FixModel';
@@ -54,7 +55,20 @@ ava.serial('.findFixByName() returns null if a FixModel does not exist within th
     t.true(result === null);
 });
 
-ava.serial('.findRealFixes() returns a list of fixes that dont have `_` prepedning thier name', t => {
+ava.serial('.getFixPositionCoordinates() returns the position of a FixModel', t => {
+    const result = FixCollection.getFixPositionCoordinates('BAKRR');
+    const expectedResult = [ 675.477318026648, -12.012221291734532 ];
+
+    t.true(_isEqual(result, expectedResult));
+});
+
+ava.serial('.getFixPositionCoordinates() returns null if a FixModel does not exist within the collection', t => {
+    const result = FixCollection.getFixPositionCoordinates('');
+
+    t.true(result === null);
+});
+
+ava.serial('.findRealFixes() returns a list of fixes that dont have `_` prepending thier name', t => {
     const result = FixCollection.findRealFixes();
 
     t.true(result.length === 104);